+++ schema/postgres/v1.9.1.sql
@@ -1,0 +1,1 @@
+ALTER TABLE "visitor_time_stats" DROP COLUMN "sessions";