+++ util_test.go
@@ -7,10 +7,9 @@ import (
 
 func TestRunAtMidnight(t *testing.T) {
 	cancel := RunAtMidnight(func() {
-		t.Fatal("Function must not be called")
+		panic("Function must not be called")
 	})
 	cancel()
-	time.Sleep(time.Millisecond * 100)
 }
 
 func TestNewTenantID(t *testing.T) {